Allow passing input to realize
This commit fixes issue #2 which raised an interesting possible use of
Operation, which is to build up a long Operation class, then realize it
multiple times with different inputs. Previously, what would require:

    $operation->input($run)->realize();

Can now be passed directly to realize:

    $operation->realize($run);

The input() method was a design choice so that operations always began
with an initial set of data and subsequent calls to the Operation
instance would show modifications being made. This style of using
Operation wasn't foreseen by me, so it's interesting to support this
possible use.

In the example given above, calling the same Operation class many times
with different inputs is now cheaper.