Allow realize via __invoke
This commit fixes the comment mentioned in #2 which called for being
able to create an Operation and call it directly via PHP's object
invoke feature.

This could be used alongside 7922c5d19c to pass around pre-built
operations that can be invoked right away, potentially without any
knowledge of the Operation class.