Cleanup documentation
This commit adds @author and @license to the class description as well
as making some minor edits to cleanup the documentation.

As there are no major changes to the documentation, and the package as
a whole was already MIT licensed, this isn't going to be tagged as 1.1,
and will just be part of the next release (likely 1.2).